chore(ui): rename file to not cause madge to fail
